+++ src/release.js
@@ -3,11 +3,11 @@ import { changelog } from './changelog'
 import { commitRelease, updateVersion } from './utils'
 
 export const release = async version => {
+  if (!version) throw new Error('Relese requires a version')
+
   const newVersion = semver.valid(semver.coerce(version))
 
-  if (!newVersion) {
-    return console.error(`Version '${version}' doesnt look right`)
-  }
+  if (!newVersion) throw new Error(`Version '${version}' doesnt look right`)
 
   try {
     await updateVersion(newVersion)